test adam for glue

+++ run_classifier.py
@@ -996,7 +996,7 @@ def model_fn_builder(electra_config, num_labels, init_checkpoint, learning_rate,
                 labels=label_ids, predictions=predictions, weights=is_real_example)
 
             mcc = MatthewsCorrelationCoefficient(num_classes=1)
-            mcc.update_state(label_ids, predictions)
+            mcc.update_state(tf.cast(label_ids, tf.float32), tf.cast(predictions, tf.float32))
             #mcc = mcc_metric(y_true=label_ids, y_pred=predictions)
             #mcc = tf.reshape(mcc, [1])
             #print(mcc)
